Fix UI jumps when hovering lines in Rapid Diffs
When line row height is not 20px it might cause a jump in UI when nearest lines are hovered because we toggle the content-visibility property for these lines.
This property relies on the intrinsic height of 20px and bigger lines will cause layout mismatch.
By adding 'auto' we ensure that the intrinsic height is automatically recalculated if the actual row height is not 20px, preventing jumps.

Changelog: fixed