Requeue backfill of slack_integrations_scopes
App scope propagation code created duplicate records
in the table. Because of this the original backfill BBM
failed. After fixing app code, we need to requeue this
migration that can handle duplicate scopes for an integration
and also that deletes these duplicates. We store deleted records
in an archive table for safety

Changelog: other