Add danger warning for e2e tests in security MRs
Regressions caused by security MRs are more difficult to resolve.
E2E test implementations may be more obscure to developers. The
purpose of this warning is to help avoid a regression caused
by misunderstanding a test, and modifying the test to pass
instead of correctly identifying an issue with application code.